+++ src/components/Capture/index.ts
@@ -1,8 +1,8 @@
-import { WithTrackingProps } from '~types/hocs'
 import { appendToTracking } from '../../Tracker'
 import Document from './Document'
 import Face from './Face'
 import ProfileData from './ProfileData'
+import ProofOfAddress from './ProofOfAddress'
 import withCaptureVariant from './withCaptureVariant'
 
 export const DocumentFrontCapture = appendToTracking(
@@ -19,10 +19,8 @@ export const DocumentVideoCapture = appendToTracking(
 )
 
 export const PoACapture = appendToTracking(
-  withCaptureVariant(Document, {
-    isPoA: true,
+  withCaptureVariant(ProofOfAddress, {
     forceCrossDevice: false,
-    side: 'front',
   }),
   'poa'
 )