+++ src/components/CountrySelector/__tests__/index.test.tsx
@@ -3,13 +3,14 @@ import { shallow } from 'enzyme'
 
 import { mockedReduxProps } from '~jest/MockedReduxProvider'
 
-import CountrySelector, { Props as CountrySelectorProps } from '../index'
 import { createOptionsSteps } from '../../Router/useOptionsSteps'
 import { NarrowSdkOptions } from '~types/commons'
 
 const defaultOptions: NarrowSdkOptions = {
   steps: [{ type: 'welcome' }],
 }
+import { Props as CountrySelectorProps } from '../index'
+import DocumentCountrySelector from '../DocumentCountrySelector'
 
 const defaultProps: CountrySelectorProps = {
   documentType: 'driving_licence',
@@ -41,7 +42,7 @@ const defaultProps: CountrySelectorProps = {
 
 describe('CountrySelector', () => {
   it('renders without crashing', () => {
-    const wrapper = shallow(<CountrySelector {...defaultProps} />)
+    const wrapper = shallow(<DocumentCountrySelector {...defaultProps} />)
     expect(wrapper.exists()).toBeTruthy()
   })
 })