+++ src/components/Router/MainRouter.tsx
@@ -16,6 +16,7 @@ import { SdkConfigurationServiceProvider } from '~contexts/useSdkConfigurationSe
 import { createOptionsSteps } from './useOptionsSteps'
 import { createWorkflowSteps } from './useWorkflowSteps'
 import { UserConsentProvider } from '~contexts/useUserConsent'
+import { PoASupportedCountriesProvider } from '~contexts/usePoASupportedCountries'
 
 const isUploadFallbackOffAndShouldUseCamera = (step: StepConfig): boolean => {
   if (!step.options || (step.type !== 'document' && step.type !== 'face')) {
@@ -44,6 +45,7 @@ export default class MainRouter extends Component<InternalRouterProps, State> {
     const {
       documentType,
       idDocumentIssuingCountry,
+      poaDocumentCountry,
       poaDocumentType,
       deviceHasCameraSupport,
       options,
@@ -83,6 +85,7 @@ export default class MainRouter extends Component<InternalRouterProps, State> {
       crossDeviceClientIntroProductName,
       crossDeviceClientIntroProductLogoSrc,
       idDocumentIssuingCountry,
+      poaDocumentCountry,
       language,
       poaDocumentType,
       step: crossDeviceInitialStep,
@@ -167,17 +170,27 @@ export default class MainRouter extends Component<InternalRouterProps, State> {
             <Spinner shouldAutoFocus={options.autoFocusOnInitialScreenTitle} />
           }
         >
-          <HistoryRouter
-            {...this.props}
-            mobileConfig={this.generateMobileConfig()}
-            onFlowChange={this.onFlowChange}
-            stepIndexType="user"
-            useSteps={
-              this.useWorkflowRun()
-                ? createWorkflowSteps(options, urls)
-                : createOptionsSteps(options)
+          <PoASupportedCountriesProvider
+            url={urls.onfido_api_url}
+            token={token}
+            fallback={
+              <Spinner
+                shouldAutoFocus={options.autoFocusOnInitialScreenTitle}
+              />
             }
-          />
+          >
+            <HistoryRouter
+              {...this.props}
+              mobileConfig={this.generateMobileConfig()}
+              onFlowChange={this.onFlowChange}
+              stepIndexType="user"
+              useSteps={
+                this.useWorkflowRun()
+                  ? createWorkflowSteps(options, urls)
+                  : createOptionsSteps(options)
+              }
+            />
+          </PoASupportedCountriesProvider>
         </UserConsentProvider>
       </SdkConfigurationServiceProvider>
     )