+++ test/utils/jest/responses.ts
@@ -26,8 +26,9 @@ export const fakePassportImageResponse: DocumentImageResponse = {
     detect_blur: { valid: true },
   },
   applicant_id: '<to-be-replaced>',
-  href: '/v3/documents/a2910652-2ed9-42d9-82b5-0e0578ab57fb',
-  download_href: '/v3/documents/a2910652-2ed9-42d9-82b5-0e0578ab57fb/download',
+  href: '/v3.3/documents/a2910652-2ed9-42d9-82b5-0e0578ab57fb',
+  download_href:
+    '/v3.3/documents/a2910652-2ed9-42d9-82b5-0e0578ab57fb/download',
 }
 
 export const fakeDrivingLicenceFrontResponse: DocumentImageResponse = {
@@ -43,8 +44,9 @@ export const fakeDrivingLicenceFrontResponse: DocumentImageResponse = {
     detect_glare: { valid: true },
   },
   applicant_id: '<to-be-replaced>',
-  href: '/v3/documents/35e2f9cb-b79a-460a-ae86-297af0dace7b',
-  download_href: '/v3/documents/35e2f9cb-b79a-460a-ae86-297af0dace7b/download',
+  href: '/v3.3/documents/35e2f9cb-b79a-460a-ae86-297af0dace7b',
+  download_href:
+    '/v3.3/documents/35e2f9cb-b79a-460a-ae86-297af0dace7b/download',
 }
 
 export const fakeDrivingLicenceBackResponse: DocumentImageResponse = {
@@ -60,8 +62,9 @@ export const fakeDrivingLicenceBackResponse: DocumentImageResponse = {
     detect_glare: { valid: true },
   },
   applicant_id: '<to-be-replaced>',
-  href: '/v3/documents/8bf3b39b-2a0e-4755-bb40-72da002d259d',
-  download_href: '/v3/documents/8bf3b39b-2a0e-4755-bb40-72da002d259d/download',
+  href: '/v3.3/documents/8bf3b39b-2a0e-4755-bb40-72da002d259d',
+  download_href:
+    '/v3.3/documents/8bf3b39b-2a0e-4755-bb40-72da002d259d/download',
 }
 
 export const fakeNoDocumentError: ParsedError = {